הוספת תגיות וארגונים - רק בעמוד האדמין Fixes #143
